+++ tenpy/linalg/np_conserved.py
@@ -853,6 +853,30 @@ class Array(object):
             extended.iset_leg_labels(labs)
         return extended
 
+    def extend(self, axis, new_ind_len, extend_charges=None):
+        """Increase the dimension of a given axis, filling the values with zeros.
+
+        Parameters
+        ----------
+        axis : int | str
+            The axis (or axis-label) to be extended.
+        new_ind_len : int
+            The new size of the specified leg.
+        charges_extend : charges | None
+            The charge values to be used for the new charge block.
+            `None` defaults to trivial charges (i.e. 0 values).
+
+        Returns
+        -------
+        extended : :class:`Array`
+            A copy of self with the specified axis increased.
+        """
+        extended = self.copy(deep=True)
+        ax = self.get_leg_index(axis)
+        extended.legs[ax] = extended.legs[ax].extend(new_ind_len, extend_charges)
+        extended._set_shape()
+        return extended
+
     # handling of charges =====================================================
 
     def gauge_total_charge(self, axis, newqtotal=None, new_qconj=None):